feat: add solution modules task
