--- src/modules/files/c.cjs
@@ -1,1 +1,0 @@
-console.log('Hello from c.cjs!');
