+++ src/modules/files/c.js
@@ -1,0 +1,1 @@
+console.log("Hello from c.js!");
